+++ ADMF/functions/Invoke-AdmfForest.ps1
@@ -126,6 +126,15 @@
 				}
 				else { Write-PSFMessage -Level Host -String 'Invoke-AdmfForest.Skipping.Test.NoConfiguration' -StringValues 'Schema (Ldif)' }
 			}
+			if ($newOptions -band [UpdateForestOptions]::NTAuthStore)
+			{
+				if (Get-FMNTAuthStore)
+				{
+					Write-PSFMessage -Level Host -String 'Invoke-AdmfForest.Executing.Invoke' -StringValues 'NTAuthStore', $parameters.Server
+					Invoke-FMNTAuthStore @parameters
+				}
+				else { Write-PSFMessage -Level Host -String 'Invoke-AdmfForest.Skipping.Test.NoConfiguration' -StringValues 'NTAuthStore' }
+			}
 		}
 		catch { throw }
 		finally { Invoke-PostCredentialProvider -ProviderName $CredentialProvider -Server $originalArgument.Server -Credential $originalArgument.Credential -Cmdlet $PSCmdlet }